+++ backend/src/main/java/com/mindstore/backend/repository/TextRepository.java
@@ -3,7 +3,7 @@ package com.mindstore.backend.repository;
 import com.mindstore.backend.data.Text;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TextRepository extends JpaRepository<Text, Long> {
+public interface TextRepository extends JpaRepository<Text, Integer> {
 
 
     boolean existsByTitle(String title);
